% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{confint.matrix}
\alias{confint.matrix}
\title{Confidence intervals for general linear hypothesis tests.}
\usage{
\method{confint}{matrix}(object, parm, level = 0.95, coef., vcov., ...)
}
\arguments{
\item{object}{numeric matrix. The (r by p) K matrix specifying the r linear
combinations for which a confidence interval should be generated.}

\item{parm}{a specification of which parameters are to be given confidence
intervals, either a vector of numbers or a vector of names. If missing, all
parameters are considered.}

\item{level}{confidence level required (default = 0.95).}

\item{coef.}{numeric vector. Parameter estimates.}

\item{vcov.}{numeric matrix. Variance-covariance matrix for the parameter
estimates.}

\item{...}{additional arguments (not currently implemented).}
}
\value{
A matrix (or vector) with columns giving lower and upper confidence
limits for each linear combination.
}
\description{
Compute asymptotic confidence intervals for linear combinations of a
parameter vector
\deqn{\mathbf{K}\beta}
}
\examples{
carsmod <- lm(dist ~ speed, data = cars)
K <- matrix(c(1, 0, 0, 1, 1, 1), nrow = 3, ncol = 2, byrow = TRUE)

confint(K, coef. = coef(carsmod), vcov. = vcov(carsmod))

}
