% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_bootstrap.R
\docType{data}
\name{.reyesEnv}
\alias{.reyesEnv}
\title{Environment for bootstrapping.}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.reyesEnv
}
\description{
Environment for bootstrapping.
}
\keyword{datasets}
