% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{KMest}
\alias{KMest}
\title{Kaplain-Meier estimates.}
\usage{
KMest(form, data, conflevel = 0.95)
}
\arguments{
\item{form}{formula for describing a \code{Surv} object.}

\item{data}{optional dataset where variables appearing in form should be
located.}

\item{conflevel}{scalar. Confidence level (default = 0.95)}
}
\value{
dataframe containing the survival estimates.
}
\description{
Compute Kaplan-Meier (product limit) estimates for survival data. This is a
wrapper to survfit in order to return a nicer set of data.
}
\examples{
library(survival)
test1 <- list(time = c(4,3,1,1,2,2,3),
              status = c(1,1,1,0,1,1,0),
              x = c(0,2,1,1,1,0,0),
              sex = c(0,0,0,0,1,1,1))

fit <- KMest(Surv(time, status) ~ sex, data = test1)

}
