% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{LifeTable}
\alias{LifeTable}
\title{Life table estimates.}
\usage{
LifeTable(nrisk, nevent, ncensor, data, conflevel = 0.95)
}
\arguments{
\item{nrisk}{vector. Number of subjects at risk at start of interval.}

\item{nevent}{vector. Number of events occuring during the interval.}

\item{ncensor}{vector. Number of subjects censored during interval.}

\item{data}{dataframe. Optional data set containing the variables.}

\item{conflevel}{scalar. Confidence level (default = 0.95).}
}
\value{
dataframe containing the survival estimates (has class LifeTable).
}
\description{
Compute life table estimates for survival data.
}
