% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_formats.R
\name{ma482_word_format}
\alias{ma482_word_format}
\title{Custom Word template.}
\usage{
ma482_word_format(...)
}
\arguments{
\item{...}{additional arguments provided to \@code{word_document}.}
}
\description{
Loads additional style file for Word for MA482.
}
