% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.LifeTable}
\alias{print.LifeTable}
\title{Print LifeTable object.}
\usage{
\method{print}{LifeTable}(x, ...)
}
\arguments{
\item{x}{LifeTable object}

\item{...}{additional arguments (not currently implemented).}
}
\description{
Prints a LifeTable object and returns it invisibly.
}
