% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{reyes482-package}
\alias{reyes482}
\alias{reyes482-package}
\title{Additional functionality for MA482/BE482 Biostatistics at RHIT.}
\description{
Additional functions useful in the analysis of data from the biological
sciences, which are not readily available in standard packages, are provided
here. These are used in the Biostatistics course at Rose-Hulman Institute of
Technology.
}
