% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{skim_clean}
\alias{skim_clean}
\title{Summarize a dataset.}
\usage{
skim_clean(data, ...)
}
\arguments{
\item{data}{see \code{\link[skimr]{skim}}}

\item{...}{see \code{\link[skimr]{skim}}.}
}
\description{
Wrapper for \code{\link[skimr]{skim}} which returns numeric summaries for
variables in a dataset.
}
