% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_plots.R
\name{stat_halfnorm}
\alias{stat_halfnorm}
\title{Halfnormal plot.}
\usage{
stat_halfnorm(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{see \code{\link[ggplot2]{stat_identity}}}

\item{data}{see \code{\link[ggplot2]{stat_identity}}}

\item{geom}{see \code{\link[ggplot2]{stat_identity}} (default = "point")}

\item{position}{see \code{\link[ggplot2]{stat_identity}} (default =
"identity")}

\item{na.rm}{boolean. If \code{TRUE}, missing values are removed (default =
FALSE)}

\item{show.legend}{see \code{\link[ggplot2]{stat_identity}}}

\item{inherit.aes}{see \code{\link[ggplot2]{stat_identity}}}

\item{...}{see \code{\link[ggplot2]{stat_identity}}}
}
\description{
Construct a half-normal plot.
}
