% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wild_bootstrap.R
\name{WildBoot}
\alias{WildBoot}
\alias{WildBoot.nls}
\alias{WildBoot.lm}
\title{Wild bootstrap.}
\usage{
WildBoot(object, f = coef, labels = names(f(object)), R = 999, ncores = 1, ...)

\method{WildBoot}{nls}(object, f = coef, labels = names(f(object)), R = 999, ncores = 1, ...)

\method{WildBoot}{lm}(object, f = coef, labels = names(f(object)), R = 999, ncores = 1, ...)
}
\arguments{
\item{object}{a regression object of class "lm" or "nls".}

\item{f}{see \code{\link[car]{Boot}}}

\item{labels}{see \code{\link[car]{Boot}}}

\item{R}{see \code{\link[car]{Boot}}}

\item{ncores}{see \code{\link[car]{Boot}}}

\item{...}{arguments passed to additional methods}
}
\value{
See \code{\link[boot]{boot}} for the returned value of the structure
returned by this function.
}
\description{
Implement a wild bootstrap for \code{lm} and \code{nls} objects. This is a
very rough approach based on the \code{\link[car]{Boot}} function.
}
\section{Methods (by class)}{
\itemize{
\item \code{nls}: 

\item \code{lm}: 
}}

\examples{
m1 <- lm(dist ~ speed, data = cars)
betahat.boot <- WildBoot(m1, R = 199)

confint(betahat.boot)

}
